+++ ec2-controller.tf
@@ -96,12 +96,12 @@ resource "aws_security_group" "controller" {
 }
 
 resource "aws_launch_template" "controller" {
-  depends_on = [ 
+  depends_on = [
     aws_iam_policy.controller
   ]
 
   name_prefix            = var.controller_instance_name
-  image_id               = data.aws_ami.ubuntu2204.id
+  image_id               = data.aws_ami.ubuntu2004.id
   instance_type          = var.controller_instance_type
   update_default_version = true