+++ main.tf
@@ -8,13 +8,13 @@ data "aws_eks_cluster_auth" "eks" {
 
 data "aws_availability_zones" "available" {}
 
-data "aws_ami" "ubuntu2204" {
+data "aws_ami" "ubuntu2004" {
   most_recent = true
   owners      = ["amazon"]
 
   filter {
     name   = "name"
-    values = ["ubuntu/images/hvm-ssd/ubuntu-jammy-22.04-amd64-server-*"]
+    values = ["ubuntu/images/hvm-ssd/ubuntu-focal-20.04-amd64-server-*"]
   }
 
   filter {